+++ cloudflare/src/lib.rs
@@ -2,6 +2,7 @@ use std::rc::Rc;
 
 use anyhow::anyhow;
 
+mod cache;
 mod env;
 mod file;
 mod handle;
@@ -17,6 +18,10 @@ pub fn init_file(env: Rc<worker::Env>, bucket_id: String) -> anyhow::Result<file
 
 pub fn init_http() -> http::CloudflareHttp {
   http::CloudflareHttp::init()
+}
+
+pub fn init_cache(env: Rc<worker::Env>) -> cache::CloudflareChronoCache {
+  cache::CloudflareChronoCache::init(env)
 }
 
 #[worker::event(fetch)]
