+++ src/cli/cache.rs
